+++ Assets/Framework/CustomGameFramework/Editor/EasyBuild/EasyBuild.Resource.cs
@@ -18,8 +18,7 @@ namespace CustomGameFramework.Editor
     {
         public static class EasyBuild_Resource
         {
-            public static async Task<bool> RunAssetBundle(BuildTarget buildTarget, string packageVersion,
-                IEncryptionServices encryptionServices, IShareAssetPackRule shareAssetPackRule, string packageName = "DefaultPackage")
+            public static async Task<bool> RunAssetBundle(BuildTarget buildTarget, string packageVersion, IEncryptionServices encryptionServices)
             {
                 // delete StreamingAssets
                 EasyBuild_Utility.DeleteFolder("Assets/StreamingAssets/BuildinFiles");
@@ -28,29 +27,37 @@ namespace CustomGameFramework.Editor
                 EasyBuild_Utility.DeleteFolder("Assets/../Bundles");
                 await EasyBuild_Utility.WaitCompile();
                 // build
-                bool isSuccess = BuildInternal(buildTarget, packageVersion, encryptionServices, shareAssetPackRule, packageName);
+                bool isSuccess = BuildInternal(buildTarget, packageVersion, encryptionServices);
                 await EasyBuild_Utility.WaitCompile();
                 return isSuccess;
             }
 
-            private static bool BuildInternal(BuildTarget buildTarget, string packageVersion,
-                IEncryptionServices encryptionServices, IShareAssetPackRule shareAssetPackRule, string packageName = "DefaultPackage")
+            private static bool BuildInternal(BuildTarget buildTarget, string packageVersion, IEncryptionServices encryptionServices)
             {
                 // 构建参数
                 string defaultOutputRoot = AssetBundleBuilderHelper.GetDefaultOutputRoot();
                 BuildParameters buildParameters = new BuildParameters();
                 buildParameters.OutputRoot = defaultOutputRoot;
                 buildParameters.BuildTarget = buildTarget;
-                buildParameters.BuildPipeline = EBuildPipeline.BuiltinBuildPipeline;
-                buildParameters.BuildMode = EBuildMode.ForceRebuild;
-                buildParameters.PackageName = packageName;
+                buildParameters.BuildPipeline = AssetBundleBuilderSettingData.Setting.BuildPipeline;
+                buildParameters.BuildMode = AssetBundleBuilderSettingData.Setting.BuildMode;
+                buildParameters.PackageName = AssetBundleBuilderSettingData.Setting.BuildPackage;
                 buildParameters.PackageVersion = packageVersion;
                 buildParameters.VerifyBuildingResult = true;
-                buildParameters.CompressOption = ECompressOption.LZ4;
-                buildParameters.OutputNameStyle = EOutputNameStyle.HashName;
-                buildParameters.CopyBuildinFileOption = ECopyBuildinFileOption.ClearAndCopyAll;
+                buildParameters.AutoAnalyzeRedundancy = true;
+                buildParameters.ShareAssetPackRule = new DefaultShareAssetPackRule();
                 buildParameters.EncryptionServices = encryptionServices;
-                buildParameters.ShareAssetPackRule = shareAssetPackRule;
+                buildParameters.CompressOption = AssetBundleBuilderSettingData.Setting.CompressOption;
+                buildParameters.OutputNameStyle = AssetBundleBuilderSettingData.Setting.OutputNameStyle;
+                buildParameters.CopyBuildinFileOption = AssetBundleBuilderSettingData.Setting.CopyBuildinFileOption;
+                buildParameters.CopyBuildinFileTags = AssetBundleBuilderSettingData.Setting.CopyBuildinFileTags;
+
+                if (AssetBundleBuilderSettingData.Setting.BuildPipeline == EBuildPipeline.ScriptableBuildPipeline)
+                {
+                    buildParameters.SBPParameters = new BuildParameters.SBPBuildParameters();
+                    buildParameters.SBPParameters.WriteLinkXML = true;
+                }
+                
                 // 执行构建
                 AssetBundleBuilder builder = new AssetBundleBuilder();
                 var buildResult = builder.Run(buildParameters);